Refactor quiz components: rename GameQuiz to GameQuizDashboard and implement MainGameQuiz for better structure and maintainability
